Stochastic GM+E Restarts v2 (#3)
In stochy_data_mod the name of the ocean stochastic restart file is no longer hard-coded. Instead it has become a public module variable initialized to its previously hard-coded value. This enables other modules to change the name of the restart file. Together with PR 376 to NCAR/MOM6, this enables MOM6 to write the restart file in the run directory, with file name equal to the mom6 restart filename but with .r_stoch. in place of .r..

It also massively reduces the amount of output printed to the cesm log file, and corrects some copy/paste typos in the output.

This partially addresses issue 360 on NCAR/MOM6 about not having exact restarts when using the stochastic GM+E parameterization.